Switch to static network config
..and other stuff 😬

Biggest change is DHCP is disabled on the patch subnets. It remains
enabled on the leaf subnets, but none of the gateway VMs should be using
DHCP themselves for network config.

Static network config is passed to the VMs as userdata. This works using
the metadata service, but as this has been unreliable on PSI I've also
enabled config drives.

Struct changes to facilitating static network config:
Ports are now stored as a struct:
  {
      id: <port uuid>,
      mac_address: <port mac address>
      ip_address: <ip address of the first fixed ip (we only expect 1)>
      subnet: <subnet uuid of the first fixed ip>
  }
Leaves in vars becomes a dict rather than a list to make it easier to
look up values corresponding to the current leaf.

Random:
Add an ICMP rull to the security group.

DNS is no longer required as a variable as we get it from the external
subnet our external port is allocated from.

Patch cidr reduce to /30 from /29 because we no longer need an IP for
DHCP. We can't reduce to /31 because neutron doesn't support RFC 3021.