leafs: sysctl: disable send_redirects
`send_redirects` should enabled only if the leaf router is the only host providing routing in the network/subnet.
However, we have custom routing rules on each host where the VIP lives
(ip rule & ip route, in a specific table), so the packets from the VIP
leaving the host are forced to go through the gateway.

For that reason, we don't want the router to send redirects and let the
router takes care of handling the packets for the VIP.

It's a complicated scenario because the VIP is not part of the same
subnet, but shares the same L2 segment; so we have to do routing (L3)
for packets that come in and go out.